Some potential bug fixes

+++ champions/src/main/java/me/mykindos/betterpvp/champions/champions/skills/skills/knight/passives/Cleave.java
@@ -58,7 +58,7 @@ public class Cleave extends Skill implements PassiveSkill, Listener {
         return percentageOfDamage + ((level - 1) * percentageOfDamageIncreasePerLevel);
     }
 
-    public double getMaxEnemiesHit(int level) {
+    public int getMaxEnemiesHit(int level) {
         return maxEnemiesHit + ((level - 1) * maxEnemiesHitIncreasePerLevel);
     }
